Add files via upload
This program is designed to find the maximum difference when there are multiple values present. In the code itself, I have used a random number generator to create a set of values to set up the structure of the code, the random numbers will be replaced when dealing with a specific set of numbers.